修复virtualbox无法同步bios日期的bug fix the bug that the bios date of a virtualbox vm cannot be synced
修复virtualbox中timeoffest变成小数的bug
fix the bug that timeOffset in virtualbox would be a double

update README.md

+++ VMGuide.Core/vbox.cs
@@ -55,7 +55,7 @@ namespace VMGuide
             set
             {
                 if (Path == null || IsLocked) return;
-                var TimeOffset = (value - DateTime.Now).TotalMilliseconds;
+                var TimeOffset =(int)((value - DateTime.Now).TotalMilliseconds);
                 XML.WriteAttribute(@"VirtualBox/Machine/Hardware/BIOS/TimeOffset", "value", TimeOffset.ToString());
             }
         }
